Update policies page UI with capability for adding a new policy
